Rm extra indirection by making (color-identifiers:refontify) an alias
For Emacs 25.1 and higher that makes `(color-identifiers:refontify)`
an alias to `(font-lock-flush)`. `(color-identifiers:refontify)` always
was just a wrapper, so no functional change is intended besides
removing the unnecessary indirection.

I haven't found a way to just leave `(if (fboundp … (defalias …)`
expression at the top-level and not cause any byte-compilation errors,
so the solution required a little bit of hackery, but hopefully it
should be safe. At least tests are passing and I haven't found any
problems.